Autotune data refactor
