+++ app/src/main/java/info/nightscout/androidaps/plugins/TuneProfile/AutotunePrep/Meal.java
@@ -3,7 +3,7 @@ package info.nightscout.androidaps.plugins.TuneProfile.AutotunePrep;
 import java.util.ArrayList;
 import java.util.List;
 
-import info.nightscout.androidaps.plugins.TuneProfile.Opts;
+import info.nightscout.androidaps.plugins.TuneProfile.data.Opts;
 import info.nightscout.androidaps.plugins.treatments.Treatment;
 
 public class Meal {
