+++ app/src/main/java/info/nightscout/androidaps/plugins/TuneProfile/AutotunePrep/Prep.java
@@ -19,10 +19,12 @@ import java.util.List;
 import info.nightscout.androidaps.data.IobTotal;
 import info.nightscout.androidaps.data.Profile;
 import info.nightscout.androidaps.db.BgReading;
-import info.nightscout.androidaps.plugins.TuneProfile.BGDatum;
-import info.nightscout.androidaps.plugins.TuneProfile.CRDatum;
-import info.nightscout.androidaps.plugins.TuneProfile.Opts;
-import info.nightscout.androidaps.plugins.TuneProfile.PrepOutput;
+import info.nightscout.androidaps.db.TemporaryBasal;
+import info.nightscout.androidaps.plugins.TuneProfile.data.BGDatum;
+import info.nightscout.androidaps.plugins.TuneProfile.data.CRDatum;
+import info.nightscout.androidaps.plugins.TuneProfile.data.IobInputs;
+import info.nightscout.androidaps.plugins.TuneProfile.data.Opts;
+import info.nightscout.androidaps.plugins.TuneProfile.data.PrepOutput;
 import info.nightscout.androidaps.plugins.TuneProfile.TuneProfilePlugin;
 import info.nightscout.androidaps.plugins.configBuilder.ProfileFunctions;
 import info.nightscout.androidaps.plugins.iob.iobCobCalculator.IobCobCalculatorPlugin;
