+++ app/src/main/java/info/nightscout/androidaps/plugins/TuneProfile/data/PrepOutput.java
@@ -1,4 +1,4 @@
-package info.nightscout.androidaps.plugins.TuneProfile;
+package info.nightscout.androidaps.plugins.TuneProfile.data;
 
 import org.json.JSONArray;
 import org.json.JSONException;
@@ -8,6 +8,8 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
+import info.nightscout.androidaps.plugins.TuneProfile.data.BGDatum;
+import info.nightscout.androidaps.plugins.TuneProfile.data.CRDatum;
 import info.nightscout.androidaps.utils.DateUtil;
 
 public class PrepOutput {
